add basic database methods
