+++ Cinema Reservation/models.py
@@ -4,19 +4,11 @@ from sqlalchemy import (Column, Integer, String, Boolean,
 from sqlalchemy.orm import relationship
 from sqlalchemy.ext.declarative import declarative_base
 
-from email.utils import parseaddr
-from re import compile, match
 
 Base = declarative_base()
 
 
 class User(Base):
-    # def __init__(self, id, username, email, password):
-    #     self.id = id
-    #     self.username = username
-    #     self.email = email
-    #     self.password = password
-
     __tablename__ = 'user'
     id = Column(Integer, primary_key=True)
     username = Column(String, unique=True)
@@ -25,34 +17,11 @@ class User(Base):
     salt = Column(String)
     superuser = Column(Boolean)
 
-    @staticmethod
-    def validate_email(email):
-        if not '@' in parseaddr(email)[1]:
-            raise ValueError('invalid email address')
-
-    @staticmethod
-    def validate_password(password):
-        if len(password) > 7:
-            has_capital_letter = compile('[A-Z]')
-            has_special_symbol = compile(r'[\W\S\D]')
-
-            if not match(has_capital_letter, password):
-                raise ValueError('password must contain a capital letter')
-
-            if not match(has_special_symbol, password):
-                raise ValueError('password must contain a special symbol')
-        else:
-            raise ValueError('password must be min 8 characters long')
-
     def __str__(self):
         return f'{self.id} | {self.username} | {self.email} | {self.password}'
 
 
 class Movie(Base):
-    # def __init__(self, id, name, rating):
-    #     self.id = id
-    #     self.name = name
-    #     self.rating = rating
     __tablename__ = 'movie'
     __table_args__ = (
         CheckConstraint('rating between 0 and 10'),
@@ -63,12 +32,6 @@ class Movie(Base):
 
 
 class Projection(Base):
-    # def __init__(self, id, movie_id, type, date, time):
-    #     self.id = id
-    #     self.movie_id = movie_id
-    #     self.type = type
-    #     self.date = date
-    #     self.time = time
     __tablename__ = 'projection'
     id = Column(Integer, primary_key=True)
     movie_id = Column(Integer, ForeignKey(Movie.id))
