+++ Cinema Reservation/setup_database.py
@@ -5,7 +5,6 @@ def setup_database():
     db = Database()
     db.create()
     # db.insert()
-    db.close()
 
 
 if __name__ == '__main__':
